fix(niri): update screenshot script

+++ pkgs/cave-screenshot.nix
@@ -4,48 +4,46 @@ pkgs.writeShellApplication {
   name = "cave-screenshot";
 
   runtimeInputs = with pkgs; [
+    grim
+    slurp
     libnotify
-    grimblast
-    hyprland
     satty
-    jq
     xdg-user-dirs
+    niri
   ];
 
   text = ''
-    border_rounding=$(hyprctl getoption decoration:rounding -j | jq .int)
-
-    # disable border roundings
-    hyprctl keyword decoration:rounding 0
-
     screenshot_name="$(date +'screenshot_%Y-%m-%d_%H-%M-%S.png')"
     satty_name="$(date +'satty_%Y-%m-%d_%H-%M-%S.png')"
     screenshot_path="$(xdg-user-dir PICTURES)/$screenshot_name"
+    timeout_time_notif="2500"
 
-    timeout_time="2500"
+    if [[ "$1" == "area" ]]; then
+        screenshot="$(slurp || true)"
+    elif [[ "$1" == "window" ]]; then
 
-    grimblast copysave "$1" "$screenshot_path" || hyprctl keyword decoration:rounding "$border_rounding"
-
-    timeout=5
-    count=0
-    while [[ ! -f "$screenshot_path" && $count -lt $timeout ]]; do
-        sleep 0.2
-        ((count++))
-    done
-
-    hyprctl keyword decoration:rounding "$border_rounding"
-
-    if [[ -f "$screenshot_path" ]]; then
-        ACTION=$(notify-send --transient -t "$timeout_time" --icon "$screenshot_path" "Screenshot saved" "Image saved in $screenshot_path" --action Edit --action Delete)
-    else
-        echo "$RESULT"
-        notify-send --transient -t "$timeout_time" --icon=dialog-error "Screenshot canceled" "$RESULT"
+        # no option to disable notification yet + only way to disable corners for screenshot is to use build-in function
+        niri msg action screenshot-window --path "$screenshot_path"
+        exit 0
     fi
 
-    if [[ "$ACTION" == "0" ]]; then
-        satty -f "$screenshot_path" -o "$satty_name"
-    elif [[ "$ACTION" == "1" ]]; then
-        rm "$screenshot_path"
+    if [[ -z "$screenshot" ]]; then
+        notify-send --transient -t "$timeout_time_notif" --icon=dialog-error "Screenshot canceled"
+        exit 0
+    else
+        grim -g "$screenshot" "$screenshot_path"
+
+        if [[ -f "$screenshot_path" ]]; then
+            ACTION=$(notify-send --transient -t "$timeout_time_notif" --icon "$screenshot_path" "Screenshot saved" "Image saved in $screenshot_path" --action Edit --action Delete)
+        else
+            notify-send --transient -t "$timeout_time_notif" --icon=dialog-error "Something went wrong"
+        fi
+
+        if [[ "$ACTION" == "0" ]]; then
+            satty -f "$screenshot_path" -o "$satty_name"
+        elif [[ "$ACTION" == "1" ]]; then
+            rm "$screenshot_path"
+        fi
     fi
   '';
 }
